+++ index.ts
@@ -90,7 +90,7 @@ const helmet: Helmet = Object.assign(
         )
       ) {
         throw new Error(
-          "Helmet no longer supports `true` as a middleware option, exception is Origin-Agent-Cluster. Remove the property from your options to fix this error."
+          "Helmet no longer supports `true` as a middleware option, except for Origin-Agent-Cluster. Remove the property from your options to fix this error."
         );
       }
     }